refactor: refactored run_all_scrapers to include a worker queue and limited the number of threads
Because Selenium is a heavy process, while running these scrapers on normal desktop/laptops we need to keep the thread count low to 5. I commented out the previous version of run_all_scrapers and included a revised version of run_all_scrapers() and a worker() fn as well. It should work logic-wise, but I am getting some stack trace errors related to chrome driver again, so we may need to revisit the UUID challenge